requirements.txt
tensorflow==2.20.0
streamlit==1.54.0
numpy==2.2.6
matplotlib==3.10.8
pillow==12.1.1
scipy==1.15.3
keras==3.12.1
